refactor(search): extract SearchResult into drivers::result
Move the live SearchResult type out of drivers::astar (which otherwise
holds only the dead legacy A* scaffold) into its own module. Behavior-
neutral; astar.rs temporarily re-exports it for the legacy test path.
