+++ crates/bloqade-lanes-search/src/drivers/mod.rs
@@ -23,3 +23,4 @@
 pub mod astar;
 pub mod entropy;
 pub mod frontier;
+pub mod result;
